% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{DGP}
\alias{DGP}
\title{Random walk.}
\usage{
DGP(n, niter)
}
\arguments{
\item{n}{sample size. Number of rows in the generated matrix.}

\item{niter}{number of columns in the generated matrix.}
}
\description{
Generate a random walk with drift 1/n.
}
\examples{
DGP(n = 100, niter = 10)
}
\keyword{generation.}
\keyword{random}
\keyword{walk}
