% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_data.R
\docType{data}
\name{sp_data}
\alias{sp_data}
\title{S&P 500 data.}
\format{A vector with the S&P 500 price dividend ratio.}
\description{
the S&P 500 price dividend ratio from January 1871 to December 2010.
}
\references{
Phillips, P.C. & Shi, S. & Yu, J. (2015a). "Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500". \emph{SSRN Electronic Journal}.
}
\keyword{dataset}
